@@ -4,7 +4,7 @@ publish_to: "none"
 version: 0.0.5
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
+  sdk: ">=2.16.0 <3.0.0"
 
 dependencies:
   animate_do: ^2.0.0
