@@ -26,13 +26,18 @@ class ChainSelectCubit extends Cubit<BlockchainState> {
   final Logger _logger;
 
   void _init() async {
-    final blockchain = _storage.getFirstItem<Blockchain>(
-      StoreKeys.Blockchain.name,
-    );
-    if (blockchain.hasError) return;
-
-    emit(BlockchainState(blockchain: blockchain.result!));
-    await Future.delayed(const Duration(milliseconds: 50));
+    final blockchain =
+        _storage.getFirstItem<Blockchain>(StoreKeys.Blockchain.name);
+    if (blockchain.hasError) {
+      if (blockchain.error! == 'empty') {
+        emit(const BlockchainState(blockchain: Blockchain.testNet));
+        await Future.delayed(const Duration(milliseconds: 50));
+      } else
+        return;
+    } else {
+      emit(BlockchainState(blockchain: blockchain.result!));
+      await Future.delayed(const Duration(milliseconds: 50));
+    }
   }
 
   void changeBlockchain(Blockchain blockchain) async {
