@@ -42,10 +42,12 @@ class SeedImportWalletState with _$SeedImportWalletState {
     return false;
   }
 
-  double completePercent() => currentStep.index / SeedImportWalletSteps.values.length;
+  double completePercent() =>
+      currentStep.index / SeedImportWalletSteps.values.length;
 
   String completePercentLabel() =>
-      ((currentStep.index / SeedImportWalletSteps.values.length) * 100).toStringAsFixed(0);
+      ((currentStep.index / SeedImportWalletSteps.values.length) * 100)
+          .toStringAsFixed(0);
 
   String currentStepLabel() {
     switch (currentStep) {
@@ -163,37 +165,30 @@ class SeedImportWalletCubit extends Cubit<SeedImportWalletState> {
     final istate = _importCubit.state;
     final wallet = istate.wallet;
     if (wallet == null) return;
+    final fullXPrv =
+        '[${wallet.fingerPrint}/${wallet.hardenedPath}]${wallet.xprv}';
 
-    final policy = 'pk([${wallet.fingerPrint}/${wallet.hardenedPath}]${wallet.xprv}/0/*)'
-        .replaceFirst('/m', '');
+    final fullXPub =
+        '[${wallet.fingerPrint}/${wallet.hardenedPath}]${wallet.xpub}';
 
-    final com = _core.compile(
+    final policy = 'pk([$fullXPrv/*)'.replaceFirst('/m', '');
+
+    const readable = 'pk(___primary___)';
+
+    final descriptor = _core.compile(
       policy: policy,
       scriptType: 'wpkh',
-    );
-
-    final exportWallet = _core.deriveHardened(
-      masterXPriv: istate.masterXpriv!,
-      account: '',
-      purpose: '92',
     );
 
     // public descriptor
 
     var newWallet = Wallet(
       label: state.walletLabel,
-      walletType: 'SINGLE SIGNATURE',
-      mainWallet: InternalWallet(
-        xPub: wallet.xpub,
-        fingerPrint: wallet.fingerPrint,
-        path: wallet.hardenedPath,
-        descriptor: com.descriptor.split('#')[0],
-      ),
-      exportWallet: InternalWallet(
-        xPub: exportWallet.xpub,
-        fingerPrint: exportWallet.fingerPrint,
-        path: exportWallet.hardenedPath,
-      ),
+      walletType: 'SIGNER',
+      descriptor: descriptor,
+      policy: readable,
+      requiredPolicyElements: 1,
+      policyElements: ['primary:$fullXPub'],
       blockchain: _blockchainCubit.state.blockchain.name,
     );
 
