@@ -6,7 +6,7 @@ part of 'wallet.dart';
 // TypeAdapterGenerator
 // **************************************************************************
 
-class WalletClassAdaper extends TypeAdapter<_$_Wallet> {
+class WalletClassAdapter extends TypeAdapter<_$_Wallet> {
   @override
   final int typeId = 1;
 
@@ -18,38 +18,41 @@ class WalletClassAdaper extends TypeAdapter<_$_Wallet> {
     };
     return _$_Wallet(
       label: fields[0] as String,
-      mainWallet: fields[1] as InternalWallet,
-      exportWallet: fields[2] as InternalWallet,
-      backedupWallet: fields[3] as InternalWallet?,
-      blockchain: fields[4] as String,
-      transactions: (fields[5] as List?)?.cast<Transaction>(),
-      id: fields[6] as int?,
-      balance: fields[7] as int?,
-      walletType: fields[8] as String,
+      descriptor: fields[1] as String,
+      policy: fields[2] as String,
+      requiredPolicyElements: fields[3] as int,
+      policyElements: (fields[4] as List).cast<String>(),
+      blockchain: fields[5] as String,
+      transactions: (fields[6] as List?)?.cast<Transaction>(),
+      id: fields[7] as int?,
+      balance: fields[8] as int?,
+      walletType: fields[9] as String,
     );
   }
 
   @override
   void write(BinaryWriter writer, _$_Wallet obj) {
     writer
-      ..writeByte(9)
+      ..writeByte(10)
       ..writeByte(0)
       ..write(obj.label)
       ..writeByte(1)
-      ..write(obj.mainWallet)
+      ..write(obj.descriptor)
       ..writeByte(2)
-      ..write(obj.exportWallet)
+      ..write(obj.policy)
       ..writeByte(3)
-      ..write(obj.backedupWallet)
+      ..write(obj.requiredPolicyElements)
       ..writeByte(4)
-      ..write(obj.blockchain)
+      ..write(obj.policyElements)
       ..writeByte(5)
-      ..write(obj.transactions)
+      ..write(obj.blockchain)
       ..writeByte(6)
-      ..write(obj.id)
+      ..write(obj.transactions)
       ..writeByte(7)
-      ..write(obj.balance)
+      ..write(obj.id)
       ..writeByte(8)
+      ..write(obj.balance)
+      ..writeByte(9)
       ..write(obj.walletType);
   }
 
@@ -59,56 +62,7 @@ class WalletClassAdaper extends TypeAdapter<_$_Wallet> {
   @override
   bool operator ==(Object other) =>
       identical(this, other) ||
-      other is WalletClassAdaper &&
-          runtimeType == other.runtimeType &&
-          typeId == other.typeId;
-}
-
-class InternalWalletClassAdaper extends TypeAdapter<_$_InternalWallet> {
-  @override
-  final int typeId = 2;
-
-  @override
-  _$_InternalWallet read(BinaryReader reader) {
-    final numOfFields = reader.readByte();
-    final fields = <int, dynamic>{
-      for (int i = 0; i < numOfFields; i++) reader.readByte(): reader.read(),
-    };
-    return _$_InternalWallet(
-      xPriv: fields[0] as String?,
-      xPub: fields[1] as String,
-      fingerPrint: fields[2] as String,
-      path: fields[3] as String,
-      descriptor: fields[4] as String?,
-      rescueDate: fields[5] as DateTime?,
-    );
-  }
-
-  @override
-  void write(BinaryWriter writer, _$_InternalWallet obj) {
-    writer
-      ..writeByte(6)
-      ..writeByte(0)
-      ..write(obj.xPriv)
-      ..writeByte(1)
-      ..write(obj.xPub)
-      ..writeByte(2)
-      ..write(obj.fingerPrint)
-      ..writeByte(3)
-      ..write(obj.path)
-      ..writeByte(4)
-      ..write(obj.descriptor)
-      ..writeByte(5)
-      ..write(obj.rescueDate);
-  }
-
-  @override
-  int get hashCode => typeId.hashCode;
-
-  @override
-  bool operator ==(Object other) =>
-      identical(this, other) ||
-      other is InternalWalletClassAdaper &&
+      other is WalletClassAdapter &&
           runtimeType == other.runtimeType &&
           typeId == other.typeId;
 }
@@ -119,14 +73,12 @@ class InternalWalletClassAdaper extends TypeAdapter<_$_InternalWallet> {
 
 _$_Wallet _$$_WalletFromJson(Map<String, dynamic> json) => _$_Wallet(
       label: json['label'] as String,
-      mainWallet:
-          InternalWallet.fromJson(json['mainWallet'] as Map<String, dynamic>),
-      exportWallet:
-          InternalWallet.fromJson(json['exportWallet'] as Map<String, dynamic>),
-      backedupWallet: json['backedupWallet'] == null
-          ? null
-          : InternalWallet.fromJson(
-              json['backedupWallet'] as Map<String, dynamic>),
+      descriptor: json['descriptor'] as String,
+      policy: json['policy'] as String,
+      requiredPolicyElements: json['requiredPolicyElements'] as int,
+      policyElements: (json['policyElements'] as List<dynamic>)
+          .map((e) => e as String)
+          .toList(),
       blockchain: json['blockchain'] as String,
       transactions: (json['transactions'] as List<dynamic>?)
           ?.map((e) => Transaction.fromJson(e as Map<String, dynamic>))
@@ -138,34 +90,13 @@ _$_Wallet _$$_WalletFromJson(Map<String, dynamic> json) => _$_Wallet(
 
 Map<String, dynamic> _$$_WalletToJson(_$_Wallet instance) => <String, dynamic>{
       'label': instance.label,
-      'mainWallet': instance.mainWallet,
-      'exportWallet': instance.exportWallet,
-      'backedupWallet': instance.backedupWallet,
+      'descriptor': instance.descriptor,
+      'policy': instance.policy,
+      'requiredPolicyElements': instance.requiredPolicyElements,
+      'policyElements': instance.policyElements,
       'blockchain': instance.blockchain,
       'transactions': instance.transactions,
       'id': instance.id,
       'balance': instance.balance,
       'walletType': instance.walletType,
-    };
-
-_$_InternalWallet _$$_InternalWalletFromJson(Map<String, dynamic> json) =>
-    _$_InternalWallet(
-      xPriv: json['xPriv'] as String?,
-      xPub: json['xPub'] as String,
-      fingerPrint: json['fingerPrint'] as String,
-      path: json['path'] as String,
-      descriptor: json['descriptor'] as String?,
-      rescueDate: json['rescueDate'] == null
-          ? null
-          : DateTime.parse(json['rescueDate'] as String),
-    );
-
-Map<String, dynamic> _$$_InternalWalletToJson(_$_InternalWallet instance) =>
-    <String, dynamic>{
-      'xPriv': instance.xPriv,
-      'xPub': instance.xPub,
-      'fingerPrint': instance.fingerPrint,
-      'path': instance.path,
-      'descriptor': instance.descriptor,
-      'rescueDate': instance.rescueDate?.toIso8601String(),
     };
